// frontend/src/api/model/user.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { Email } from './email'
import type { Name } from './name'
import type { PhoneNumber } from './phoneNumber'
import type { Residence } from './residence'
import type { UserRole } from './userRole'
import type { UserStatusProperty } from './userStatusProperty'

export interface User {
  /** The ID of the user */
  id: string
  /** The first name of the user */
  name: Name
  /** The last name of the user */
  surname: Name
  /** The phone number of the user */
  phone: PhoneNumber
  /** The email of the user */
  email: Email
  /** The hashed password of the user */
  passwordHash: string
  /** The role of the user */
  role: UserRole
  /** The residence information of the user */
  residence: Residence
  /** The status of the user */
  status: UserStatusProperty
}