// frontend/src/api/model/userProductDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { Currency } from './currency'
import type { ModelName } from './modelName'
import type { SerialNumber } from './serialNumber'
import type { UserProductDtoStatus } from './userProductDtoStatus'

export interface UserProductDto {
  /** Product ID */
  id: string
  /** Product serial number */
  serialNumber: SerialNumber
  /** Product model name */
  modelName: ModelName
  /** Product model ID */
  modelId: string
  /** Product status */
  status: UserProductDtoStatus
  /** Product price */
  salePrice: Currency
}