// frontend/src/api/model/createUserDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { Country } from './country'
import type { FederativeUnit } from './federativeUnit'

export interface CreateUserDto {
  /**
   * User name
   * @minLength 2
   * @maxLength 50
   */
  name: string
  /**
   * User last name
   * @minLength 2
   * @maxLength 50
   */
  surname: string
  /** User phone number */
  phone: string
  /** User email */
  email: string
  /**
   * User password
   * @minLength 8
   */
  password: string
  /** Street address */
  street: string
  /** Address number */
  number: number
  /**
   * Address complement
   * @nullable
   */
  complement?: string | null
  /** Neighborhood */
  neighborhood: string
  /** City */
  city: string
  /** Federative unit */
  federativeUnit: FederativeUnit
  /** Postal code */
  postalCode: string
  /** Country */
  country: Country
}