// frontend/src/api/model/role.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */

/**
 * The new role for the user
 */
export type Role = (typeof Role)[keyof typeof Role]

export const Role = {
  ADMIN: 'ADMIN',
  RESELLER: 'RESELLER',
  ASSISTANT: 'ASSISTANT',
  UNASSIGNED: 'UNASSIGNED'
} as const